ReorderList.java
Given a Singly Linked List a0 -> a1 -> a2 -> .. -> a(n-1) -> an. The task is to reorder it to a0 -> an -> a1 -> a(n-1) -> a2 -> a(n-2) -> ..

Input
First line contains a single integer N - the size of the list
Second line contains N space separated integers - the nodes in the list (Insertion at beginning)

Constraints
1 <= N <= 10000
1 <= X <= 10000 where X = value of node

Output
Print the list after reordering.