BinaryRightSideView.java
Right View of Binary Tree

Given a Binary Tree, find Right view of it. Right view of a Binary Tree is set of nodes visible when tree is viewed from right side.

Input format:
The input consists of one line, space-separated elements denoting the nodes of the Binary Tree.

Output format:
Print the right view of the tree in a single line.