ValidateBinarySearchTree.java
Check if a Binary Tree is BST or not

A binary search tree (BST) is a node based binary tree data structure which has the following properties. 

The left subtree of a node contains only nodes with keys less than the node’s key.
The right subtree of a node contains only nodes with keys greater than the node’s key.
Both the left and right subtrees must also be binary search trees.
Write a program to accept the nodes of a Binary tree. Construct a Binary tree in level order fashion and check if it is a Binary Search Tree or not.

Note:
You are to create a Binary Tree first from a parent array in level-order fashion. That is, elements from left in the array will be filled in the tree level wise starting from level 0.

For example, if the array is: {1, 2, 3, 4, 5, 6}
The structure of the tree is:
                  1
                 / \
                2   3
               / \ /
              4  5 6

Input format:
The input consists of two lines.
The first line contains an integer n, the number of nodes in the Binary tree.
The second line contains n space-separated integers, the keys for the Binary tree.

Output format:
Print True if the given tree is a BST, else print False.