PrimsAlgorithms.java
Prims MST
A connected and undirected graph is given. Find the cost of minimum spanning tree.

Note : If there is no edge given for a node, you may skip that node in MST.

Input Format:
First line - n and m, the number of nodes and edges
Then there are m lines each containing u, v, and w.
u and v are the nodes and w is the weight of the edge.

Output Format:
A single integer representing the cost of minimum spanning tree.